backend/requirements.txt
# ===== PhasQ Backend Layers (Render.com Deployment) =====

# Core
fastapi>=0.115.0  # Upgrade for Python 3.13 support
uvicorn[standard]>=0.30.0
pydantic>=2.9.0   # Upgrade for Python 3.13 support
python-multipart>=0.0.9 # Required for OAuth2 forms

# HTTP & Auth
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing
numpy==1.26.3
python-dotenv==1.0.0

# Google Earth Engine (ENABLED for Render.com - higher memory limit)
earthengine-api>=1.0.0

# NOTE: Render.com Free tier has 512MB RAM - sufficient for GEE
# These were removed for Vercel (250MB limit) but work fine on Render

